migrate: move benefitsHubLinks to components directory
- Migrated benefitsHubLinks component files to src/components/benefitsHubLinks/
- Moved and renamed index.tsx -> template.tsx
- Moved and renamed index.test.tsx -> template.test.tsx
- Moved benefitsHubLinks.ts query file to query.ts
- Moved benefitsHub.ts formatted type file to formatted-type.ts
- Moved benefitHubs.mock.json to mock.json
- Moved benefitsHubLinks.test.tsx to query.test.ts
- Moved benefitsHubLinks test snapshot to __snapshots__/query.test.ts.snap
- Updated import statements across codebase:
  - src/data/queries/index.ts updated query import
  - src/components/resourcesSupport/template.tsx updated template import
  - src/components/resourcesSupport/formatted-type.ts updated type import
- Fixed internal imports:
  - template.test.tsx: ./template instead of ./index
  - template.tsx: ./formatted-type instead of @/types/formatted/benefitsHub
  - query.ts: ./formatted-type instead of @/types/formatted/benefitsHub
  - query.test.ts: ./mock.json and ./query instead of external paths
- Removed empty src/templates/common/benefitsHubLinks/ directory
- All TypeScript compilation checks passing
- All tests passing (1 template test, 2 query tests with 1 snapshot)

This benefitsHubLinks component renders a section of VA benefits links with analytics tracking.
It includes query logic for fetching landing page support services and formatting them as links.
