+++ src/components/benefitsHubLinks/query.test.ts
@@ -4,9 +4,9 @@
 
 import { NodeLandingPage } from '@/types/drupal/node'
 import { queries } from '@/data/queries'
-import mockData from '@/mocks/benefitHubs.mock.json'
+import mockData from './mock.json'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
-import { params } from '../benefitsHubLinks'
+import { params } from './query'
 
 // field_related_office is causing issues here, I believe because the referenced node is unpublished (node/38439)
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
