+++ src/components/resourcesSupport/template.tsx
@@ -5,7 +5,7 @@ import { Button as FormattedButton } from '@/types/formatted/button'
 import { Paragraph } from '@/templates/components/paragraph'
 import { AudienceTopics } from '@/templates/components/audienceTopics'
 import { Button } from '@/templates/common/button'
-import { BenefitsHubLinks } from '@/templates/common/benefitsHubLinks'
+import { BenefitsHubLinks } from '@/components/benefitsHubLinks/template'
 import { ContactInfo } from '@/templates/components/contactInfo'
 import { RateYourExperience } from '@/templates/components/rateYourExperience'
