+++ src/data/queries/index.ts
@@ -6,7 +6,7 @@ import * as AlertNonReusable from '@/data/queries/alertNonReusable'
 import * as AlertSingle from '@/data/queries/alertSingle'
 import * as AudienceTopics from '@/data/queries/audienceTopics'
 import * as Banners from '@/components/banner/query'
-import * as BenefitsHub from '@/data/queries/benefitsHubLinks'
+import * as BenefitsHub from '@/components/benefitsHubLinks/query'
 import * as Button from '@/data/queries/button'
 import * as CollapsiblePanel from '@/data/queries/collapsiblePanel'
 import * as CollapsiblePanelItem from '@/data/queries/collapsiblePanelItem'
